app/requirements.txt
cassandra-driver
venv-pack
pyarrow
fastparquet
pathvalidate
wheel
pyspark
tqdm